moved useClickOutside back to Datepicker
