+++ packages/asc-ui/src/index.ts
@@ -119,7 +119,6 @@ import { perceivedLoading } from './utils/themeUtils'
 
 export {
   useDebounce,
-  useClickOutside,
   useActionOnEscape,
   useEdgeDetection,
   useFocus,
